requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.10
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-filter==0.2.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
exceptiongroup==1.2.0
hiredis==2.3.2
idna==3.6
iniconfig==2.0.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
pycparser==2.21
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.5.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
svgwrite==1.4.3
tomli==2.0.1
Tree==0.2.4
tree-cli==0.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
